pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "healthlink"
version = "1.0.0-mvp"
description = "HealthLink - AI-driven Comorbidity Management Platform"
authors = [{name = "HealthLink Team"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web框架
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    
    # 数据库
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pymysql>=1.1.0",
    
    # 数据验证
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # HTTP客户端
    "httpx>=0.25.0",
    
    # 安全
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    
    # 配置
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    
    # 日志和监控
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",
    
    # 工具
    "click>=8.1.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    # 代码质量
    "black>=23.11.0",
    "isort>=5.12.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    
    # 测试
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    
    # 开发工具
    "pre-commit>=3.5.0",
]

[project.scripts]
healthlink = "healthlink.cli:main"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "W", "C90", "I", "N", "UP", "B", "A", "C4", "T20"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]